update2 in task-12

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -5,7 +5,7 @@
  */
 void print_number(int n)
 {
-	long power, longN;
+	long long power, longN;
 	int check;
 
 	longN = n;
